+++ economicsl/account.py
@@ -1,0 +1,39 @@
+class Account:
+    def __init__(self, name, accountType, startingBalance=0.0) -> None:
+        self.name = name
+        self.accountType = accountType
+        self.balance = np.longdouble(startingBalance)
+
+    # A Debit is a positive change for ASSET and EXPENSES accounts, and negative for the rest.
+    def debit(self, amount):
+        if (self.accountType == AccountType.ASSET) or (self.accountType == AccountType.EXPENSES):
+            self.balance += amount
+        else:
+            self.balance -= amount
+
+    # A Credit is a negative change for ASSET and EXPENSES accounts, and positive for the rest.
+    def credit(self, amount):
+        if ((self.accountType == AccountType.ASSET) or (self.accountType == AccountType.EXPENSES)):
+            self.balance -= amount
+        else:
+            self.balance += amount
+
+    def getAccountType(self):
+        return self.accountType
+
+    def getBalance(self):
+        return self.balance
+
+    def getName(self) -> str:
+        return self.name
+
+
+def enum(**enums):
+    return type('Enum', (), enums)
+
+
+AccountType = enum(ASSET=1,
+                   LIABILITY=2,
+                   INCOME=4,
+                   EXPENSES=5,
+                   GOOD=6)
