+++ economicsl/bankersrounding.py
@@ -1,0 +1,14 @@
+import numpy as np
+
+class BankersRounding:
+    def bankersRounding(self, value: np.longdouble) -> int:
+        s = int(value)
+        t = abs(value - s)
+
+        if ((t < 0.5) or (t == 0.5 and s % 2 == 0)):
+            return s
+        else:
+            if (value < 0):
+                return s - 1
+            else:
+                return s + 1
