+++ economicsl/contract.py
@@ -1,0 +1,15 @@
+class Contract:
+    def getAssetParty(self):
+        pass
+
+    def getLiabilityParty(self):
+        pass
+
+    def getValue(self, me):
+        pass
+
+    def getAvailableActions(self, me):
+        pass
+
+    def getName(self, me):
+        pass
