+++ economicsl/ledger.py
@@ -1,58 +1,3 @@
-import numpy as np
-from abce import NotEnoughGoods
-from .account import Account
-from .ledger import Ledger
-
-
-def doubleEntry(debitAccount, creditAccount, amount: np.longdouble):
-    debitAccount.debit(amount)
-    creditAccount.credit(amount)
-
-class Contracts:
-    def __init__(self):
-        self.allAssets = []
-        self.allLiabilities = []
-
-class Account:
-    def __init__(self, name, accountType, startingBalance: np.longdouble=0.0) -> None:
-        self.name = name
-        self.accountType = accountType
-        self.balance = np.longdouble(startingBalance)
-
-    # A Debit is a positive change for ASSET and EXPENSES accounts, and negative for the rest.
-    def debit(self, amount: np.longdouble):
-        if (self.accountType == AccountType.ASSET) or (self.accountType == AccountType.EXPENSES):
-            self.balance += amount
-        else:
-            self.balance -= amount
-
-    # A Credit is a negative change for ASSET and EXPENSES accounts, and positive for the rest.
-    def credit(self, amount: np.longdouble):
-        if ((self.accountType == AccountType.ASSET) or (self.accountType == AccountType.EXPENSES)):
-            self.balance -= amount
-        else:
-            self.balance += amount
-
-    def getAccountType(self):
-        return self.accountType
-
-    def getBalance(self) -> np.longdouble:
-        return self.balance
-
-    def getName(self) -> str:
-        return self.name
-
-
-def enum(**enums):
-    return type('Enum', (), enums)
-
-
-AccountType = enum(ASSET=1,
-                   LIABILITY=2,
-                   INCOME=4,
-                   EXPENSES=5,
-                   GOOD=6)
-
 
 # This is the main class implementing double entry org.economicsl.accounting. All public operations provided by this class
 # are performed as a double entry operation, i.e. a pair of (dr, cr) operations.
@@ -81,20 +26,20 @@ class Ledger:
         self.liabilityAccounts = {}  # a hashmap from a contract to a liabilityAccount
         self.me = me
 
-    def getAssetValue(self) -> np.longdouble:
+    def getAssetValue(self):
         return sum([aa.getBalance() for aa in self.assetAccounts.values()])
 
-    def getLiabilityValue(self) -> np.longdouble:
+    def getLiabilityValue(self):
         return sum([la.getBalance() for la in self.liabilityAccounts.values()])
 
-    def getEquityValue(self) -> np.longdouble:
+    def getEquityValue(self):
         return self.getAssetValue() - self.getLiabilityValue()
 
-    def getAssetValueOf(self, contractType) -> np.longdouble:
+    def getAssetValueOf(self, contractType):
         # return assetAccounts.get(contractType).getBalance();
         return sum([c.getValue() for c in self.contracts.allAssets if isinstance(c, contractType)])
 
-    def getLiabilityValueOf(self, contractType) -> np.longdouble:
+    def getLiabilityValueOf(self, contractType):
         # return liabilityAccounts.get(contractType).getBalance();
         return sum([c.getValue() for c in self.contracts.allLiabilities if isinstance(c, contractType)])
 
@@ -189,11 +134,11 @@ class Ledger:
         elif (new_value < old_value):
             self.getGoodsAccount(name).credit(old_value - new_value)
 
-    def addCash(self, amount: np.longdouble) -> None:
+    def addCash(self, amount) -> None:
         # (dr cash, cr equity)
         self.create("money", np.longdouble(amount), 1.0)
 
-    def subtractCash(self, amount: np.longdouble) -> None:
+    def subtractCash(self, amount) -> None:
         self.destroy("money", np.longdouble(amount), 1.0)
 
     # Operation to pay back a liability loan; debit liability and credit cash
